Add demo of coroutines using `async`/`await`
We've been thinking about coroutines and generators in Rust lately due
to our work on RFC 3513[^1].  While revising the RFC, we realized that
we could implement synchronous coroutines (and thereby generators) by
using (or abusing, perhaps) `async`/`await`.  We set out to create a
demonstration.  This project is that demonstration.

The key idea here is that we want Rust to synthesize a state machine
from a `async` block, but otherwise we don't want or need any of the
`Future` machinery.  We don't need an executor, and we don't need to
wake anything.  We simply drive the resulting `Future` when resuming
the coroutine.

Within a coroutine `async` block, we await invocations of `yield`.
But remember that this is actually synchronous, so trying to await
anything else is going to result in a panic or a deadlock (but should
be memory safe).

The resulting API is remarkably reasonable.  The library does no
allocations and is `no_std`.

This is our first working version.

The demonstration raises many questions that apply to the actual
article we may want in the language.

E.g., we decided to have the first call to `resume` accept a value.
This value becomes the result of the `yield` expression.  Execution
continues until another `yield` is encountered or until the `Future`
resolves, and `resume` returns either the yielded value or the yielded
value (if any) and the return value of the block.

However, another way to do it would be to require that a `start`
method be called first without arguments.  Execution would proceed
until the first yield or until the end of the block.  That method
would return either the yielded value or the return value of the
block.  Then each call to `resume` would pass in a value that would
become the result of the `yield` expression and execution would
continue until the next `yield` or until the block returns.

There are some advantages to both approaches.

[^1]: https://github.com/rust-lang/rfcs/pull/3513