Rework coroutine lifetime with `start`/`resume`
We had been starting execution of the coroutines with the first call
to `resume`.  This doesn't work out as well as one might hope.  E.g.,
if the coroutine doesn't yield at all, then the value passed in is
just silently ignored.  It also forces us, on the last call to
`resume`, to return both the final return value and the argument to
the last `yield`, if there was one.  Of course this means we need an
option just to handle the case of a coroutine that does not yield.

It makes a lot more sense to instead `start` each coroutine without an
argument.  Let's do that.  Then we can `resume` the coroutine with
arguments.  In this way, both `start` and `resume` can have more
homogeneous return values.  They return either the argument to yield
or the return value of the coroutine.

This also makes the internals of the coroutine lifecycle more
homogeneous.