Add tests demonstrating dangling pointers
We need to fix the memory unsafety we've introduced.  The first most
obvious example is that the `Yielder` can escape the closure to which
we pass it.  We'll need to do something about that.  For now, let's
add a test to demonstrate the problem.  Run with `cargo miri test` to
see the failures.