Fix unsafety in safe API
Our API allowed for UB in safe code.  We pass a `Yielder` to a
closure.  That `Yielder` references a field in the `Coro` that creates
it.  We were allowing that yielder to escape the closure with nothing
to prevent it from outliving the `Coro`.  This would result in a
dangling reference and use-after-free.

There are two aspects to this fix.  First, we need to stop the
`Yielder` from escaping the closure directly to an outer scope.  We do
this with a `Self: 's` bound on the `impl` of the `initialize` method.
This is where the bad things were happening in this case, because it's
here that we extend the lifetime of the reference in `Yielder` to
match `'s` and pass the `Yielder` to the closure.  Because `'s` could
be longer than `Self`, we were extending this lifetime too far.  By
bounding `Self: 's`, we enforce that `'s` must have exactly the same
lifetime as `Self`.  That is, since `Self` contains `'s`, `'s:
'self` (if we could say that); because of the bound, `Self: 's`.  So
they must be equal.

Secondly, we were allowing the `Yielder` to be captured by the
closure, then captured by the future, then returned by the future into
an outer scope.  The `advance` method was responsible for this.

One would almost think that with `Self: 's`, this would take care of
itself.  The `advance` method returns `Output<T, U>`.  In the
substitution, `U` is going to be the `Yielder` that was passed
through.  This type parameter should have captured the `'s` lifetime
in `initialize`.  Since we've ensured that `Self: 's`, one could be
forgiven for thinking that would be enough.

Here's another thing that doesn't work.  One might think that bounding
`U: 's` should definitely do it.  Consider:

```rust
where
  F: FnOnce(Yielder<'s, T, R>) -> G,
  G: Future<Output = U>,
  T: 's
  U: 's,
  Self: 's,
```

The future outputs `U`.  If we bound `U: 's`, and if the `Yielder` is
substituted for `U`, and if `'s` doesn't outlive the `Coro`, then one
might think this would definitely do it.

As it turns out, though, we need to pass a lifetime explicitly through
to the `Output`.  Passing through the lifetime of the `Pin<&mut Self>`
reference works fine.  Using the longer `'s` lifetime also works
fine.

The other theoretical possibility to test would be the `Yielder`
escaping by being yielded.  If this would possible, it would also be
handled by our changes to include a lifetime parameter in `Output`.
But it's not actually possible, as trying to `yield` the `Yielder`
results in a "cyclic type of infinite size".