Fix soundness holes correctly
After some analysis, we worked out how to do this correctly.  The
coroutine needs to be pinned before we initialize it, since that
initialization introduces a self-reference, so we do that now.  We use
a builder pattern to make this relatively ergonomic.

The first step of the magic happens in `CoroBuilder::init`.  We need
for the final `Coro<'s, ..` to have a lifetime `'s` that's not too
long.  To achieve this, we bound it by the lifetime of the pinned
mutable reference.  The content can't be overwritten or invalidated
while that lifetime is valid.

The second step of the magic is in `<Coro as Resumable>::initialize`.
We're called on a pinned mutable reference that may be shorter than
`'s`.  We need to allow this so that e.g. `as_mut()` works so we can
call into these methods more than once.  But we need a reference to
the state that has a lifetime of `'s` so we can store that lifetime
back in the self-referential struct.  We can do this because we
ensured that `'s` is short enough and because we're not handing out
mutable references to the state.

Also noteworthy is that it's important that `CoroBuilder::init` has
the full set of where clauses used later in the `Resumable` impl.  If
it doesn't, then as soon the future actually captures the `Yielder`
things break because Rust can't prove that the `Yielder` outlives the
future.

Getting all of this right means we can clean up other stray places we
had added `PhantomData` and lifetimes to try to make this work, so
let's do that.