Handle panics while constructing coroutine
If the provided closure were to panic while we're constructing the
coroutine, we would fail to drop the field already initialized since
the struct is not yet initialized.  This isn't much of a problem here
since we're not allocating and the field will not have a non-trivial
value at this point.  But let's follow the more correct pattern.

Note that on `no_std` this is just a no-op.