Add `Iterator` implementation
Let's implement `Iterator` on top of `Coro`.  We're not sure yet
whether this one is useful.

Because we can't do a blanket implementation for `Iterator` using
`Generator` as we would like to, due to the orphan rule, let's also
implement our own `MyIterator` trait so we can demonstrate how such a
blanket impl  would work.