Remove unneeded lifetime parameter from `Coro`
We had included a lifetime parameter in `Coro` and in `CoroBuilder`.
We were binding this lifetime parameter to the lifetime of the pinned
mutable reference to the `CoroBuilder` (and to `Coro` after `init`).
We use this as the lifetime of the self-reference that we incorporate
in the `Yielder` that is captured by the future we store in the
`Coro`.

However, we don't actually need for this to be a lifetime parameter on
`Coro` and on `CoroBuilder`.  The `CoroBuilder::init` method takes the
lifetime of the pinned mutable reference to `CoroBuilder` as a generic
argument.  That's all we need.  That's the exact lifetime we need for
the self-reference and for the lifetime of the returned pinned mutable
reference to `Coro`.

Let's remove the lifetime parameter that we don't need.