Add more soundness tests
We were testing to ensure that the `Yielder` could not be leaked and
live too long, but we also need to test that 1) a reference cannot be
fed into the coroutine and then held after the referent is dropped and
that 2) a reference cannot be yielded from the coroutine and then held
after the coroutine (or the pinned mutable reference to it) is
dropped.

Let's test to ensure that such code does not compile.