Resolve aliasing shared reference concerns
We were populating the `Yielder` with a shared reference to the state
cell.  This state cell was owned by `Coro` which is behind a mutable
reference at all relevant times.  The shared reference and the mutable
reference would have been aliasing each other.

Since the references were to a `Cell`, this did not violate the shared
XOR mutable rules of Rust.

However, is it OK to create this shared reference that aliases a
mutable reference?

Arguably it is.  The shared reference would have been `!Send` as it
pointed to a `Cell`.  And we were ensuring that the shared reference
could not outlive the mutable reference from which it was derived.
Consider that this is OK:

```rust
let mut x0 = ((),);
let x0 = &mut x0;
let x1 = &x0.0;
dbg!(x0.0);
dbg!(x1);
```

Similarly, the aliased shared reference here was arguably fine.

However, the rules for when it is allowed to create a reference are
nuanced and not yet fully defined, so let's err on the side of caution
and just use mutable raw pointers for accessing the state.  We must
still, of course, ensure that our use of aliased pointers produces
well-defined behavior.

Since we're using raw mutable pointers which are always derived from
mutable references, we no longer have interior mutability, and we no
longer need to use `Cell`.