Resolve Stacked Borrows error
We were getting an error from Stacked Borrows (but not from Tree
Borrows).  We introduced this error in the (rather benign) commit "Add
explicit tracking of coroutine start state".

We would need to think more carefully about whether (and if so, why)
this might represent a real problem.  However, in any case we can
resolve it by having the future capture only the `NonNull` pointer
rather than capturing a mutable reference to the `Self` type or to one
of its fields, so let's do that.  This is better regardless, as it
reduces what needs to be captured by the returned hidden type of
`yield`.

In reducing the captures, of course, we have to be careful that we
don't reduce them too far.  We need the returned future to not outlive
the state to which it holds a pointer.  So let's add more explicit
handling for this, and let's add the relevant tests.